pyproject.toml
[project]
name = "dteam-django-practical-test"
version = "0.1.0"
description = "Test task for DTEAM"
authors = [
    {name = "VolodymyrMamaevsky",email = "vova.moder@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "weasyprint (>=60.1,<61.0.0)",
    "openai (>=1.95.1,<2.0.0)",
    "gunicorn (>=22.0.0,<23.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.3"
mypy = "^1.16.1"
pre-commit = "^4.2.0"
django-stubs = "^5.2.1"
pytest-django = "^4.11.1"
